(#11379) Add param for ports.pp
Prior to this commit, puppetlabs-apache is not managing port listen
entries explicity and would either create duplicate listen directives
or worse, not have them at all.

This commit adds a $ports_conf parameter to reference a ports.conf
file that we'll manage with these directives.

The code comes from cwarden in pull request 7 but an alternative
pull request is being creates using the file_line function instead
of execs or concat.