Fix broken tests for ssl, vhost, vhost::*
A number of tests were broken by the last few style commits when '755'
became '0755' and the tests didn't get updated. This commit fixes the
tests so that they work again.

The tests for apache::ssl were broken because the 'operatingsystem' fact
was not being set in the tests. Don't know how long that's been busted,
but in any case this commit fixes that problem as well by adding
appropriate let(:facts).