Ensure proxy mods are loaded by vhost::proxy
At the moment, defining an apache::vhost::proxy resource will lead to
Apache failing to start by default. This is because nothing ensures the
relevant proxy modules are loaded into Apache. This patch fixes that on
Debian and Ubuntu. Hopefully, someone can fill in the blanks on the
RedHat family of operating systems.