Adding template for httpd.conf and default mods
The default apache httpd.conf contains settings that need to be
available for tweaking. Currently the puppetlabs-apache module does
not manage this configuration file, so this module captures the default
configuration that may be updated for templating in the future.

This commit also paves the way for more-heavily modularlizing the apache
mod configuration pulling mod configuration out of httpd.conf and
creating templates per mod.

This change is backwards-compatible for anyone using the
puppetlabs-apache module to manage their apache installation (ie, not
using external sources to manage httpd.conf).