Control server-status location configuration Without this commit, the configuration of the location, server-status used with the status module is not controllable.
This commit breaks the config out into a status.conf file. This is done
by controlling the status module and the configuration file in a new.
class called apache::mod::status. To remain consistent with previous
behavior of this module, it is included by default and defaults to
server-status being disabled.