Remove enable parameter from apache::mod::info
Originally, apache::mod::info was designed with the ability to be
included but disabled. This was a misinterpretation of default apache
behavior. Ultimately, it is not required to have an enabled option.

This patch removes the 'enable' parameter from apache::mod::info. It will
now be dependent on the declaration of the class as to wether or not
mod_info is active.