(#266) Allow apache::mod to specify explicit module identifier and module path
Using the Shibboleth Apache module as an example, this is the default
way in which the pre-built, RPM-packaged shibboleth install wants to
be called by an Apache config file:

LoadModule mod_shib /opt/shibboleth-sp/lib/shibboleth/mod_shib_22.so

Under the current puppetlabs-apache setup, modules must be called
XXXX_module and the path must be under the standard module directory.

This adds a path parameter and an identifier parameter.  For now, I
use the $varname_REAL convention that was already in use in this class
definition, although this could easily be switched to the $_varname
convention used elsewhere in the puppetlabs-apache module.  In any
case, if either $path or $identifier is supplied, the config directive
looks like this:

LoadModule $identifier $path

If either is omitted, the code constructs them based on the original
logic.

Tests are included, but I have had difficulty getting tests to run.
Test verification by someone with a better setup than me would be
greatly appreciated