Remove default <proxy *> block
This is a horrible and generally useless default. Most people do *not*
use apache httpd as Forward proxy, and those who do, should explicitly
set an allow. That this is exactly what our new, and backward
compatible proxy config does, and allows, while widly simplifying the
reverse proxy configurations.

We add a test case for new proxy configuration to make sure this is
actually the case, and also fix a test case from a previious commit.