make default mods configurable
...in a backwards-compatible manner.

Right now there are two options, either to enable a rather broad batch
of modules that someone found to be a good default set, or to leave them
unmanaged all-together.

This patch provides the users wtih the ability to manage their set of
default modules, by passing it as array to default_mods.

This fixes #298

n.b.: we've added test cases here, but they aren't yet part of the
actual test spec, because I have no idea how to do that.