added $server_root parameter
Previously, the $httpd_dir was used as ServerRoot in httpd.conf
template. On some installations httpd_dir is not same as ServerRoot,
for example on FreeBSD apache22 (installed via ports) has ServerRoot set
to /usr/local by default.

This PR adds $server_root parameter to apache::params. Its purpose is to
be substituted in httpd.conf as ServerRoot. This parameter is used
in #342. The purpose of the whole thing is to plit #342 into smaller
parts to make review process easier, see
https://github.com/puppetlabs/puppetlabs-apache/pull/342#issuecomment-25423813