Add mod_rpaf support
From apt-cache show libapache2-mod-rpaf:

Rpaf is short for reverse proxy add forward. It changes the remote
address of the client visible to other Apache modules when two
conditions are satisfied. First condition is that the remote client is
actually a proxy that is defined in httpd.conf. Secondly if there is an
incoming X-Forwarded-For header and the proxy is in its list of known
proxies it takes the last IP from the incoming X-Forwarded-For header
and changes the remote address of the client in the request structure.