Set SSLOptions StdEnvVars in server context
We set `SSLOptions` now in server context and we set it to `StdEnvVars`.
Thus, we no longer need the <FilesMatch > in server context to set it
for php, CGI and SSI applications.
We allow it to be overridden by the `apache::mod::ssl` class parameter
`ssl_options`. Further, we allow it to be overridden in directory
context.
Documentation for all of these has been extended and tests have been
adapted to make sense in this new context. This commit fixes #487