Remove index parameters.
The current code allows you to set index_options but then ignores it.
Based on the git log for the commit that added this option it was
intended to be passed into directories.  The test that caught this
will be merged into spec/acceptance/vhost_spec.rb in a seperate PR.

You can reproduce the original failure with:

class { 'apache': }
host { 'test.server': ip => '127.0.0.1' }
apache::vhost { 'test.server':
  docroot    => '/tmp',
  index_options    => ['Charset=UTF-8'],
}

We're just going to remove them because they can be passed directly
into directory instead of indirectly.