Configure Passenger in separate .conf file on RH so PassengerRoot isn't lost
The mod_passenger RPM supplies /etc/httpd/conf.d/passenger.conf with the
correct PassengerRoot for the version of the package installed.  The most
reliable way to keep the PassengerRoot accurate is to leave this file in
place and then to install a second file (passenger_extra.conf) for other
Passenger customisations.

Hardcoding the PassengerRoot in the module means it gets out of step with
package repositories (3.0.21 is in EPEL6 at the time of writing, while the
module assumes .17).  Passenger 4.x refuses to start if the PassengerRoot is
incorrect.  The user of apache::mod::passenger is also able to override this
value, but they shouldn't be forced to, just to get a functioning install.

Fixes #560