Allow overriding the detected $apache_name.
This will mainly be useful for supporting Red Hat Software Collections,
where 'httpd24-httpd' installs Apache 2.4 alongside Apache 2.2 in the
base repository. The $service_name parameter already allows
customization of the service to use, but not the package.

Document the ::apache::apache_name parameter.

Add a test for the apache_name param.