Add support for specifying the docroot option for RHEL SCL httpd24
The pattern of defining these variables in the init.pp goes back into 2012.
However, it's much better to use the params.pp pattern for this purpose, and
it's also required to make this work with the RHEL's Software Collections.
Without this patch, the Apache::Vhost['default'] is instantiated (regardless
whether it's with ensure => absent or present) and bails out due to /var/www
being missing on asystem which has never had RHEL6's regular httpd package
installed.