(FM-1908) fixup mod_php tests
Previously we were testing that the source attribute for
apache::mod::php overrode the content and template attributes. This
began failing with the removal of files/spec file (it was the target
used by the source attribute in testing. These are the kinds of tests
that are slated for removal with the test refactor, best course is to
remove them now.