Removed enforced .conf extension on custom_config.pp
The current custom_config automatically updates the file to use .conf
rather than rely on the name provided. Any configs/vhosts/etc. using
this have to predict this behavior rather than rely on what's provided
in the name of the config.

This is not exposed or configurable so removing it should help make
things easier from a configuration aspect.