Created custom_config extension override
The custom_config resource forcibly renamed the file to include .conf as
the extension. The behavior was not documented and did not reflect the
name of the resource in this case. Any configs including those would
need to remember to include .config even if the base file name did not
have it.

By adding the parameter and defaulting it to '.conf', this commit allows
backward compatibility while also allowing custom config extensions in
the future.