Ensure module packages are installed before evaluating conf.d dir
The problem I'm running into on EL7 is that when the ssl module is
enabled, the mod_ssl package gets installed and drops
`/etc/httpd/conf.d/ssl.conf`. We want to remove that file, but the purge
for `/etc/httpd/conf.d` is being evaluated before the mod_ssl package is
being installed. On the second client run the `ssl.conf` file will be
removed.

This change ensures that the `mod_ssl` package gets installed prior to
the purge on `/etc/httpd/conf.d` being evaluated so that the ssl.conf
file gets removed on the first client run.