Add JkMount/JkUnmount directives to vhost
'mod_jk' is one of the possibilities to use Apache as a frontend to
Tomcat. The module understands 'JkMount' and 'JkUnMount' directives to
configure whether an URL should be handled by Apache or by Tomcat. This
patch enhances 'apache::vhost' to allow configuration of these two
directives for a virtual host.