vhost: add support for <Limit> block in directories
In order to allow "require" statements inside the <Limit> blocks without
introducing code duplication, the template for "require" handling has
beenrefactored into a subtemplate of its own. This subtemplate accesses
the current block being processed (be it a <Directory> or <Limit>)
(ab)using a scope variable _template_scope, which is in turn a hash with
a single variable :item, to workaround puppet's "thou shall not reset
variables".