MODULES-5297: Move vhost error_log and log_level into incremental defined type
Create the defined type `apache::vhost::logging` to define the
error_log and log_level configuration of an apache vhost.  Only one
such type may be specified per vhost.

If the `error_log` parameter is true, or if a `log_level` parameter
is passed to the `apache::vhost` resource, it automatically creates
a corresponding `apache::vhost::logging` resource.