(maint) apache unit test failure
I'm making an educated guess that the error log message for puppet 4.7.1 is formatted in a different way from 4.10.4. the end part of the message, "got 'foo'", may not be present in 4.7.1 or is in a different location. this change still captures the important part of the message while allowing tests to pass.