Replace legacy `bool2httpd()` function with shim
This commit replaces the legacy
`lib/puppet/parser/functions/bool2httpd.rb` function with a 4.x API shim
function that calls the namespaced `apache::bool2httpd` function.

The shim might not actually be necessary as the function is probably
`private`.

The non-namespaced legacy version was still being used in a few
instances in templates.  This has been corrected.