Remove OS fact differences in proxy_connect_spec
Using `it_behaves_like` creates a new scope and runs those tests in them.
`a mod class, including apache` is a shared context and doesn't contain
any tests. When using `include_examples`, you can actually inherit the
context.

It then goes a step further and reuses that context in every test. This
works because the Apache version is a parameter so it doesn't matter.

After that, nothing refers to the Debian 7 facts anymore and can be
cleaned up.