(MODULES-10899) Install PHP 7 by default on SLES >= 15
On SLES >= 15 it appears as though the `apache2-mod_php5` package
is no longer available in the default repo.

This commit ensures that the default value for the $php_version
parameter is 7 on SLES >= 15 and the appropriate `apache2-mod_php7`
package is used.

The spec test expectations have also been updated based on the
changes.