Only warn about servername logging if relevant
The warning about $use_servername_for_filenames is long and may not
affect users. If they use a normal $name, there would effectively be no
difference. This compares $name to $normalized_servername. The warning
is only relevant Only if they differ.

Technically there could be a space in $servername (since there is no
data type to enforce validation) but Apache wouldn't accept that anyway.

It also fixes two typos in the warnings.