Change the v8 vhost filename logging logic
Previously the plan was to change $use_port_for_filenames to always true
in v8. However, that creates a major upgrade pain. It is better to only
do so if it's non-default.

This also changes the upgrade warning logic to match.

I'd also like to emphasize that this is a really painful plan. The
current logging really works a lot better and I'd still like to propose
to revert this whole plan for v8.