Remove warnings and plans to change vhost default naming
In 7e233fe7cc108605384506f11083aeea6e3a2464 a plan was made to change
the naming. In 6b2a20a390df597f04222571dc7fac366a8ea898 it was already
from 6.0.0 to 7.0.0. Now that 7.0.0 has been released, the default isn't
going to change. This time I'm suggestion to drop the plan altogether
and stop bothering users with this.