Acceptance tests: Migrate to puppet/epel
stahnma/epel was transferred to puppet/epel. However, version 4.0.0
drops EL6 support (while adding EL9 support) so a conditional is added.
The API is the same so the later code can be used on both. This
conditional limits module installation to Red Hat-based OSes which makes
sure it truly is a soft dependency on other operating systems.

It also removes the EPEL setup in mod_authnz_ldap tests since EPEL
should always be available after it's being set up in the acceptance
helper.