Tighten apache::vhost::directories parameter
This tightens the data type to a simple array of hashes (as the
parameter documentation suggests). This allows the logic to be simpler.
Because of this, it is now fairly easy to include the correct modules if
needed. Some options are implemented but there are more which could be
automatically included.