Declare mod_php unsupported on EL9
EL 9 doesn't ship mod_php, so this is uninstallable. By making it a
required parameter (which effectively was already the case) and setting
the version to undef the catalog fails to compile by default. This still
allows users to set it if they build a mod_php somehow.

EL 9 users should use php-fpm instead.

Fixes: edbdb65b06418779aa30ebe6cf0948007cc51a29