Manage DNF module for mod_auth_openidc
On EL 8 mod_auth_openidc is in a DNF module that must be enabled.
Otherwise the package is uninstallable. This is verified by adding an
acceptance test for the class.

The inheritance on apache::params is removed since it was redundant.
That is only needed if a class parameter uses apache::params.

$oidc_settings on apache::vhost is changed to have a default. The
template expects one and With that it's impossible to miscompile. The
alternative is to implement a fail() inside the code if it is empty, but
this provides some safety.