(maint) Add variable manage_vhost_enable_dir
Additional variable added at request of community to allow for `vhost_enable_dir` to be disabled on Debian systems.
Due to the way in which aid variable inherits from the params manifest, simply setting it to `undef` does not work as when this is done it's logic assumes it should inherit from the default params value.