fix rewrite rules being ignored
In 7b22dae the default value for $rewrites was changed
from undef to [] in manifests/vhost.pp. This results
in templates_vhost/_rewrite.erb always evaluating to
false (boolean check on an empty array variable is
still true) unless $rewrites is explicitly set to undef.
This results in the content of $rewrite_rule being ignored.

closes: #2318