Extract acceptance node setup to a separate file
This makes it easier to maintain. You get proper syntax highlighting in
your editor and can even lint it. This also converts some shell lines to
pure Puppet. This reduces the need for large local
spec_helper_acceptance modifications. The pattern is copied from
voxpupuli-acceptance where the beaker setup automatically detects if the
file is present and applies it.

It does drop EL6 EPEL support, following
4922544709313e0b7d176013e6f8118b85fdf96a.