Stop managing mime support package on Debian
All apache2 packages on Debian 10 and later already have a hard dependency on
the `mime-support` package. While `mime-support` still exists on Debian 12
as a transitonal package, upcoming Debian 13 has removed it entirely.

This sets the `$mime_support_package` to undefined so that the package
isn't managed on Debian, since it doesn't need to be.