Stop using apache::apache_version
In cedd45b63be89ea54bd2a596e6cd3a3f60d4faf8 this was dropped and the
whole module is now only for Apache 2.4+. This means the parameter is
redundant. The parameter is kept in so this can be released in a bugfix
release, but the next major version should drop it.

Fixes: cedd45b63be89ea54bd2a596e6cd3a3f60d4faf8 ("Drop Apache 2.2 support")