Install mod_http2 on EL if required
In EL8 httpd had a hard requirement on mod_http2 but in EL9 it turned
into a weak dependency. While most installs default to installing weak
dependencies, it can be disabled by users. OracleLinux even defaults to
disabled.

EL7 doesn't have http2 at all so the whole class can't be used with it.